requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
idna==3.11
kiwisolver==1.4.5
matplotlib==3.8.4
mysql-connector-python==9.5.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.16.0
tabulate==0.9.0
urllib3==2.6.2
